Implement save/restore still_ins
This is to allow for recovering from a round where everyone
gets eliminated.  Rather than just start over, I can back up
on round, to the last set of still-ins, and play another round
with those folks.

Also use constants for STILL_IN and OUT values.

Also, make is so that forms submitted for the wrong question number
or wrong round, or that are late (after the answer or result is shown)
are rejected and the player is eliminated, with a warning message.