Make the default game mode 'group'
Eliminate the 'suppress_refresh' file, so that a normal
admin-driven game can occur.  Change the game mode in rg.cgi
to 'group'.

This is the default configuration for a multiplayer game.
(which still works, as far as I can tell).

Also, finalize a set of example questions in rgdata/trivia.py