make it so edit_game and set_game_data don't raise an exception
The 'edit_game' feature doesn't work right now.  For some reason,
when you save the game data is has no effect.  But calling
it would raise an exception because it tried to access items that
were not hashable, or (on set) try to set items that were not game
data.  Get rid of the exceptions by only processing elements of
the data instance that are game data.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>