Add some auxiliary files (start_server, etc.)
Add start_server, which can be used for a python2 test cgiserver.
Also, add the file 'one-unicode-problems', which is a user file
that includes characters that cause problems for Python3.  This was
a user file (for the 'one' account) which was used for testing
the conversion from python2 to python3 for the rg.cgi script.